requirements.txt
aws-cdk.aws-sns
aws-cdk.aws-lambda
aws-cdk.aws-lambda-event-sources
